+++ squadcastv1/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [typescript v0.2.0] squadcastv1
 ### Releases
 - [NPM v0.2.0] https://www.npmjs.com/package/@solarwinds/squadcast-sdk-typescript/v/0.2.0 - squadcastv1
+
+## 2025-10-23 00:30:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.638.1 (2.729.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.1] squadcastv1
+### Releases
+- [NPM v0.2.1] https://www.npmjs.com/package/@solarwinds/squadcast-sdk-typescript/v/0.2.1 - squadcastv1